+++ packages/typespec-go/test/local/armtest/zz_pets_client_example_test.go
@@ -66,7 +66,7 @@ func ExamplePetsClient_AdoptPet() {
 	_ = res
 	// If the HTTP response code is 200 as defined in example definition, your response structure would look as follows. Please pay attention that all the values in the output are fake values for just demo purposes.
 	// res = armtest.PetsClientAdoptPetResponse{
-	// 	Dog: armtest.Dog{
+	// 	PetClassification: &armtest.Dog{
 	// 		Kind: to.Ptr(armtest.PetKindDog),
 	// 		Name: to.Ptr("Buddy"),
 	// 		Breed: to.Ptr("Labrador"),